+++ src/warray_wval.h
@@ -18,9 +18,9 @@
 def_struct ( warray ) {
   wtype * type;
   wtype * warray_elem_type;
-  size_t space;
-  size_t past_end;
-  size_t start;
+  long space;
+  long past_end;
+  long start;
   wval * data;
   };
 
@@ -38,10 +38,10 @@ wdeclare_base ( warray );
 #define wval_cast_kind_pointer( expr ) ( ( wobj * ) ( expr ) )
 
 bool warray_init ( warray * self, wtype * warray_elem_type, wcall * error );
-bool warray_init_to_size ( warray * self, wtype * warray_elem_type, size_t num_elements, wcall * error );
+bool warray_init_to_size ( warray * self, wtype * warray_elem_type, long num_elements, wcall * error );
 
 warray * warray_new ( wtype * warray_elem_type, wcall * error );
-warray * warray_new_to_size ( wtype * warray_elem_type, size_t num_elements, wcall * error );
+warray * warray_new_to_size ( wtype * warray_elem_type, long num_elements, wcall * error );
 
 void warray_deinit ( warray * self );
 void warray_delete ( warray * self );
@@ -54,11 +54,11 @@ void warray_push_front ( warray * self, wval elem, wcall * error );
 wval warray_pop_back ( warray * self, wcall * error );
 wval warray_pop_front ( warray * self, wcall * error );
 
-size_t warray_length ( warray * self );
-bool warray_good_index ( warray * self, size_t index );
-wval * warray_index ( warray * self, size_t index );
-wval warray_get ( warray * self, size_t index );
-wval warray_set ( warray * self, size_t index, wval val );
+long warray_length ( warray * self );
+bool warray_good_index ( warray * self, long index );
+wval * warray_index ( warray * self, long index );
+wval warray_get ( warray * self, long index );
+wval warray_set ( warray * self, long index, wval val );
 
 warray_iter warray_start ( warray * parent );
 warray_iter warray_end ( warray * parent );
